Remove visualisation layers when VisualisationScene is enabled
Visualisation layers should have their lifespan pinned to the lifetime
of a connection. If visualisation layers live longer than a connection,
then the molecule from a previous connection remains visible until the
new connection sends frame (issue #58); also we take the risk of the
rendering preferences from the previous connection to remain in the new
one.

This commit removes all the existing visualisation layers when the
VisualisationScene is enabled (which happens when the client connects to
a simulation). This way, when the same OnEnable method creates the base
layer, the base layer is indeed the first layer. It therefore avoids
issues with the base layer being duplicated (issue #66).

This is an alternative to !123.

Fixes #58
Fixes #66