Update default force scaling factor to 2500
This change is to be merged with the "feature/imd_fixes" branch on Narupa Protocol.  The default force scaling factor has been increased from 100 to 2500 (the mid-point of the scale).  This is for two reasons:
(1) The mass-weighting fix means that the same scaling factor now 'feels' less strong, so the default value has been increased to account for this. 
(2) It makes sense to have a default value at the mid-point of the scale.