Use a submodule for the NarupaUnityPlugin
Historically, the NarupaUnityPlugin was developed in the Narupa iMD
repository and automatically copied for each merge in Narupa iMD into
its own repository. This was because, despite being occasionally used in
other projects, the plugin was mostly used by Narupa iMD and it was more
convenient to develop it there. Also, new comers to the lab tended to
have issues setting up the repository with the submodule.

With the NarupaUnityPlugin used and developed more independently from
Narupa iMD, the historical process makes less sense. This commit
replaces the NarupaUnityPlugin files by a git submodule.