Fix wrong commit reference in submodule
PR #1 initially linked to the gitlab repository rather than the github
one. While the PR fixes the reference to the repository, it did not
change the reference to the specific commit causing an issue when
fetching the submodule as the github repository did not have a commit
with the referenced hash.