Copy the SteamVrControllerDefinitionTests from the plugin
The plugin defines a series of test to assert that the host project has the required assets to support some
predefined controllers (Index, Oculus, Vive, and Fallback). This test, by its very nature, depends on files
from the host project, meaning it will necesseraly fail when run against the plugin itself.

Here, we copy the test in the NarupaIMD project so it can run here, where all the required files are, rather
than in the plugin. Ultimately, it is the responsibility of the host to choose what controllers it want to
support.